Save network file data locally
Update tests to match expectations from previous changes to favorting and notes.

+++ core/network/src/main/java/com/crisiscleanup/core/network/model/NetworkFile.kt
@@ -11,7 +11,7 @@ data class NetworkFile(
     val blogUrl: String? = null,
     @SerialName("created_at")
     val createdAt: Instant,
-    val file: Int? = null,
+    val file: Long? = null,
     @SerialName("file_type_t")
     val fileTypeT: String,
     @SerialName("filename")
